Recoloured tiles, better ball animation.
Added a frame to the ball animation and redrew it as a rotating
square so it's more visible.

Changed ball shadow colour to black, rather than dim ball colour,
to make it more visible.

Changed tile colours in font to be dim player colour, to make ball
more visible when passing over tiles.