Thrust harvesting now works in the current direction.
Don't have to steer to collect thrust, and when you apply it,
it goes in the direction you are going.

Tuned square aging and harvesting to be at more reasonable
rates, and not leave behind dregs when harvesting, and show
more of the animation stages of the tiles (slowed down harvest
speed for that).